% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjClustBand.R
\name{adjClustBand}
\alias{adjClustBand}
\title{Adjacency-constrained clustering of a band diagonal similarity matrix}
\usage{
adjClustBand(x, p, h, flavor = c("crayons", "Koskas"), minNbBlocks = 1,
  verbose = FALSE)
}
\arguments{
\item{x}{A vector of length \eqn{p*h - h*(h+1)/2} similarities}

\item{p}{An integer, the number of items to be clustered}

\item{h}{An integer, the size of the diagonal band to be considered}

\item{flavor}{A character vector, the type of algorithm to be performed. Defaults to 'crayons' as described in Dehman (2015)}

\item{minNbBlocks}{An integer, the number of blocks at which clustering should be stopped (to save time). Defaults to 1, which corresponds to perform the clustering until all items are merged.}

\item{verbose}{A logical value: should extra information be displayed?}
}
\description{
Adjacency-constrained clustering of a band diagonal similarity matrix
}
\examples{

data("R2.100", package="adjclust")
x <- R2.100

h <- 100
p <- 603
res <- adjClustBand(x, p, h)
resK <- adjClustBand(x, p, h, flavor="Koskas")
}
\references{
Dehman A. (2015). "Spatial Clustering of Linkage Disequilibrium blocks for Genome-Wide Association Studies". PhD thesis. \url{https://tel.archives-ouvertes.fr/tel-01288568/}
}

