% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjClustBand_heap.R
\name{adjClustBand_heap}
\alias{adjClustBand_heap}
\title{Constrained Hierarchical Agglomerative Clustering}
\usage{
adjClustBand_heap(mat, h, blMin = 1, verbose = FALSE)
}
\arguments{
\item{mat}{A similarity matrix or a dist object}

\item{h}{band width. It is assumed that the similarity between two items is 
0 when these items are at a distance of more than band width h}

\item{blMin}{depth of clustering. It is number of clusters at which the algorithm stops. Default value is 1.}

\item{verbose}{Currently not used}
}
\value{
Function \code{adjClustBand_heap} returns an object of
class \code{\link[stats]{hclust}}.
}
\description{
Function to perform adjacency-constrained hierarchical agglomerative clustering
}
\details{
\code{adjClustBand_heap} performes constrained hierarchichal agglomerative clustering which is   
hierarchical agglomerative clustering in which each observation is associated to a position, 
and the clustering is constrained so as only adjacent clusters are merged. These methods are  
useful in various application fields, including ecology (Quaternary data) and bioinformatics 
(e.g. in Genome-Wide Association Studies (GWAS))
}
\examples{
sim <- matrix(c(1,0.1,0.2,0.3,0.1,1,0.4,0.5,0.2,0.4,1,0.6,0.3,0.5,0.6,1), nrow=4)
h <- 3
fit1 <- adjclust:::adjClustBand_heap(sim, h,1,FALSE)
plot(fit1)

dist <- as.dist(sqrt(2-(2*sim)))

#Compatibility with dist objects
fit2 <- adjclust:::adjClustBand_heap(dist, h,1,FALSE)
plot(fit2)

}
