% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chac.R
\name{chac}
\alias{chac}
\alias{as.hclust.chac}
\alias{print.chac}
\alias{head.chac}
\alias{summary.chac}
\alias{plot.chac}
\alias{diagnose.chac}
\title{Methods for class 'chac'}
\usage{
\method{as.hclust}{chac}(x, ...)

\method{print}{chac}(x, ...)

\method{head}{chac}(x, ...)

\method{summary}{chac}(object, ...)

\method{plot}{chac}(x, y, ...)

\method{diagnose}{chac}(x, graph = TRUE)
}
\arguments{
\item{x, object}{an object of class 'chac'}

\item{...}{for \code{\link{plot}}, arguments passed to the function 
\code{\link[stats]{plot.dendrogram}}. Default values for \code{type} and
\code{leaflab} are respectively set to \code{"triangle"} and \code{"none"}}

\item{y}{not used}

\item{graph}{(logical) whether the diagnostic plot has to be displayed or 
not. Default to \code{TRUE}}
}
\description{
Methods for class 'chac'
}
\details{
\code{\link{diagnose}} invisibly exports a data frame with the 
numbers of non increasing merges described by the labels of the clusters 
being merged at this step and at the previous one, as well as the 
corresponding merge heights.
}
