% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_statistic.R
\name{gap_statistic}
\alias{gap_statistic}
\title{Compute GAP statistic for each cluster number}
\usage{
gap_statistic(hcl, k_array, p.value = 0.001, mc.cores = 1)
}
\arguments{
\item{hcl}{hierarchical clustering result, where \code{hcl$data} is a similarity matrix}

\item{k_array}{array storing the number of clusters where the GAP statistic is evaluated.  All missing integer within \code{range(k_array)} are estimated using a linear interpretation of the GAP statistic.}

\item{p.value}{p-value cutoff used to determine the number of clusters}

\item{mc.cores}{number of CPU cores to use.}
}
\value{
\itemize{
  \item df_approx - data.frame storing results for each value of k
  \item n_clusters - number of clusters use p-value cutoff
}
}
\description{
Compute GAP statistic for each cluster number
}
