% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSim.R
\name{ggPlotSim}
\alias{ggPlotSim}
\title{Plot (dis)similarity matrix}
\usage{
ggPlotSim(
  mat,
  type = c("similarity", "dissimilarity"),
  clustering = NULL,
  log = TRUE,
  legendName = "intensity",
  main = NULL,
  priorCount = 0.5,
  stats = c("R.squared", "D.prime"),
  h = NULL,
  axis = FALSE,
  naxis = 10,
  axistext = NULL,
  xlab = "objects",
  cluster_col = "darkred"
)
}
\arguments{
\item{mat}{matrix to plot. It can be of class \code{'matrix'}, 
\code{'dgCMatrix'}, \code{'dsCMatrix'}, \code{'dist'}, \code{'HTCexp'},
\code{'snpMatrix'}.}

\item{type}{input matrix type. Can be either \code{"similarity"} or 
\code{"dissimilarity"} (kernels are supposed to be of type 
\code{"similarity"}).}

\item{log}{logical. Should the breaks be based on log-scaled values of the
matrix entries. Default to \code{TRUE}.}

\item{legendName}{character. Title of the legend. Default to 
\code{"intensity"}.}

\item{main}{character. Title of the plot. Default to \code{NULL} (no title).}

\item{priorCount}{numeric. Average count to be added to each entry of the
matrix to avoid taking log of zero. Used only if \code{log = TRUE}.}
}
\description{
Heatmap of the (dis)similarity matrix
}
\details{
This function produces a heatmap for the used (dis)similarity matrix that 
can be used as a diagnostic plot to check the consistency between the 
obtained clustering and the original (dis)similarity
}
\seealso{
\code{\link{select}}, \code{\link{adjClust}}
}
