% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic_imr90_40_XX.R
\docType{data}
\name{hic_imr90_40_XX}
\alias{hic_imr90_40_XX}
\title{Example Hi-C contact map}
\format{An object of class HTCexp}
\source{
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE35156
}
\usage{
hic_imr90_40_XX
}
\description{
An HTCexp object named "hic_imr90_40_XX" containing a HI-C contact map for 
chromosome X vs chromosome X, obtained from the 'HiTC' package
}
\details{
The script to create this object is available in the 'system'
  directory of the 'adjclust' package.
}
\keyword{datasets}
