% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic_imr90_40$chrXchrX.R
\docType{data}
\name{obj}
\alias{obj}
\title{The HTCexp object named "obj" contains intrachromosmal contact map of chromosome X with itself}
\format{An object of class HTCexp.}
\source{
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE35156
}
\usage{
obj
}
\description{
An HTCexp object named "obj" is obtained from HiC dataset published by Dixon et al (GSE35156) that contains all genome-wide contact maps.
}
\keyword{datasets}
