% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSim.R
\name{plotSim}
\alias{plotSim}
\title{Plot (dis)similarity matrix}
\usage{
plotSim(mat, type = c("similarity", "dissimilarity"), palette = heat.colors,
  breaks = 10, log = TRUE, h = p - 1, stats = c("R.squared", "D.prime"),
  main = NULL)
}
\arguments{
\item{mat}{matrix to plot. It can be of class \code{'matrix'}, 
\code{'dgCMatrix'}, \code{'dsCMatrix'}, \code{'dist'}, \code{'HTCexp'},
\code{'snpMatrix'}.}

\item{type}{input matrix type. Can be either \code{"similarity"} or 
\code{"dissimilarity"} (kernels are supposed to be of type 
\code{"similarity"}).}

\item{palette}{color palette. Default to \code{\link[grDevices]{heat.colors}}}

\item{breaks}{number of breaks used to set colors from the palette. Those
are based on the quantiles of the matrix entries and for skewed distributions
the actual number used to set the palette can be lower than \code{breaks}.}

\item{log}{logical. Should the breaks be based on log-scaled values of the
matrix entries. Default to \code{TRUE}.}

\item{h}{if \code{mat} is of class \code{"snpMatrix"}, band parameter used to
compute the linkage desiquilibrium (see \code{\link[snpStats]{ld}}).}

\item{stats}{if \code{mat} is of class \code{"snpMatrix"}, type of linkage
desiquilibrium measure (see \code{\link[snpStats]{ld}}).}

\item{main}{graphic title.}
}
\description{
Heatmap of the (dis)similarity matrix
}
\details{
This function produces a heatmap for the used (dis)similarity matrix that 
can be used as a diagnostic plot to check the consistency between the 
obtained clustering and the original (dis)similarity
}
\examples{
# input as HiTC::HTCexp object
if (require("HiTC", quietly = TRUE)) {
  load(system.file("extdata", "hic_imr90_40_XX.rda", package = "adjclust"))
  plotSim(hic_imr90_40_XX)
}
plotSim(dist(iris[ ,1:4]), log = FALSE)
}
