% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevfit.R
\docType{data}
\name{prevfit}
\alias{prevfit}
\title{Results obtained from primitive version of adjClustBand_heap function}
\format{An object of class hclust.}
\source{
primitive version of adjClustBand_heap function.
}
\usage{
prevfit
}
\description{
A hclust object obtained from primitve version of adjClustBand_heap function
}
\keyword{datasets}
