% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.100.R
\docType{data}
\name{R2.100}
\alias{R2.100}
\title{Linkage disequilibrium on chromosome 22 in the European population}
\format{A sparse matrix of class \code{\link[Matrix]{dgCMatrix-class}}.}
\source{
Data extracted from \code{\link[snpStats]{ld.example}}. The original data is
  from the HapMap project.
}
\usage{
R2.100
}
\description{
A dataset containing R-squared linkage disequilibrium (LD) statistics for
\eqn{p=603} SNPs spanning a one megabase regions on chromosome 22, in a
sample of 90 Europeans (CEPH). LD values were only calculated for entries in
a diagonal band of size \code{h=100}; the other values are assumed to be 0.
}
\keyword{datasets}

