% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_adjclust_0.3.0.R
\docType{data}
\name{res_adjclust_0.3.0}
\alias{res_adjclust_0.3.0}
\title{Results obtained from a primitive version of the package}
\format{An object of class 'hclust'.}
\source{
adjClust v <= 0.3.0. See script 'inst/system/create_prevfit.R'
}
\usage{
res_adjclust_0.3.0
}
\description{
An 'hclust' object obtained from a primitive version of the package
}
\keyword{datasets}
