% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chac.R
\name{select}
\alias{select}
\alias{select.chac}
\title{Clustering selection}
\usage{
select(x, k.max = NULL, pct = 0.15, graph = FALSE)
}
\arguments{
\item{x}{an object of class 'chac'}

\item{k.max}{maximum number of clusters that can be selected. Default to 
\code{NULL}, in which case it is set to 
\eqn{\min(\max(100, \frac{n}{\log(n)}), \frac{n}{2})} where \eqn{n} is the
number of objects to be clustered}

\item{pct}{minimum percentage of points for the plateau selection in 
capushe selection. See \code{\link[capushe]{DDSE}} for further details}

\item{graph}{logical. Whether the diagnostic plot for the capushe selection
is displayed or not. Default to \code{FALSE}}
}
\value{
The function returns the clustering selected by the slope heuristic,
as implemented in the R package \code{capushe}.
}
\description{
Clustering selection from a chac object with the slope heuristic
}
\examples{
if (require("HiTC", quietly = TRUE)) {
  load(system.file("extdata", "hic_imr90_40_XX.rda", package = "adjclust"))
  log_hic <- log(intdata(hic_imr90_40_XX) + 1)
  log_hic <- as(log_hic, "matrix")
  res <- adjClust(log_hic)
  selected.clust <- select(res)
  table(selected.clust)
}
}
\references{
Baudy, J.P., Maugis, C. and Michel, B. (2012) Slope heuristics: 
overview and implementation. \emph{Statistics and Computing}, \strong{22}(2),
355-470.
}
