% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmatrix.R
\docType{data}
\name{simmatrix}
\alias{simmatrix}
\title{Similarity matrix used in proving equivalence of full band adjclust and rioja}
\format{A sparse matrix of class base::matrix.}
\source{
This matrix is generated by scaling the matrix provided in medium tests.
}
\usage{
simmatrix
}
\description{
A 150 X 150 similarity symmetric numeric matrix without missing values.
}
\keyword{datasets}
