% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpClust.R
\name{snpClust}
\alias{snpClust}
\title{Constrained Hierarchical Agglomerative Clustering of Single Nucleotide 
Polymorphisms}
\usage{
snpClust(x, h = ncol(x) - 1, ...)
}
\arguments{
\item{x}{either a genotype matrix of class snpStats::SnpMatrix/base::matrix 
or a linkage disequlibrium matrix of class Matrix::dgCMatrix}

\item{h}{band width. If not provided, \code{h} is set to default value `p-1`
where `p` is the number of columns of `x`}

\item{\dots}{Further arguments to be passed to the \code{snpStats::ld}
function}
}
\value{
Function \code{snpClust} returns an object of class 
  \code{\link[stats]{hclust}}.
}
\description{
Function to perform adjacency-constrained hierarchical agglomerative 
clustering of Single Nucleotide Polymorphisms
}
\details{
\code{snpClust} performs constrained hierarchichal agglomerative clustering 
of single nucleotide polymorphisms. Constrained Hierarchical Agglomerative 
Clustering is hierarchical agglomerative clustering in which each observation
is associated to a position, and the clustering is constrained so as to merge
only adjacent clusters.
}
\examples{
## a very small example
data(testdata, package = "snpStats")

# input as snpStats::SnpMatrix
fit1 <- snpClust(Autosomes[1:200, 1:5], h = 3, stats = "R.squared")

# input as base::matrix
fit2 <- snpClust(as.matrix(Autosomes[1:200, 1:5]), h = 3, stats = "R.squared")

# input as Matrix::dgCMatrix
ld <- snpStats::ld(Autosomes[1:200, 1:5], depth = 3, stats = "R.squared")
fit3 <- snpClust(ld, 3)

}
\seealso{
\code{\link{adjClust}} \code{\link[snpStats:ld]{ld}}
}
