% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcss.R
\name{wcss}
\alias{wcss}
\title{Compute within-clstuer sum of squares}
\usage{
wcss(hcl, k_array, p.value = 0.001, mc.cores = 1)
}
\arguments{
\item{hcl}{hierarchical clustering result, where \code{hcl$data} is a similarity matrix}

\item{k_array}{array storing the number of clusters where the GAP statistic is evaluated.  All missing integer within \code{range(k_array)} are estimated using a linear interpretation of the GAP statistic.}

\item{p.value}{p-value cutoff used to determine the number of clusters}

\item{mc.cores}{number of CPU cores to use.}
}
\value{
\itemize{
  \item df_approx - data.frame storing results for each value of k
  \item n_clusters - number of clusters using p-value cutoff
}
}
\description{
Compute within-clstuer sum of squares for many values of cluster number.
}
\details{
The Within-Cluster Sum of Squares is computed for many values of cluster number, k.    Since evaluating the WCSS for each k is expensive, the WCSS for missing values of k are filled in using a linear interpolation. WCSS is used as to evaluted the number of clusters in the data, and is part of the GAP score \insertCite{tibshirani2001estimating}{adjclust}.  Here we adapted the score to be used for similarity matrices \insertCite{dehman2015performance}{adjclust}.
}
\references{
\insertRef{dehman2015performance}{adjclust}

\insertRef{tibshirani2001estimating}{adjclust}
}
