% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chac.R
\name{chac}
\alias{chac}
\alias{as.hclust.chac}
\alias{print.chac}
\alias{head.chac}
\alias{summary.chac}
\alias{plot.chac}
\alias{diagnose.chac}
\title{Methods for class 'chac'}
\usage{
\method{as.hclust}{chac}(x, ...)

\method{print}{chac}(x, ...)

\method{head}{chac}(x, ...)

\method{summary}{chac}(object, ...)

\method{plot}{chac}(x, y, ..., mode = c("standard", "corrected", "inertia"))

\method{diagnose}{chac}(x, graph = TRUE, verbose = TRUE)
}
\arguments{
\item{x, object}{an object of class 'chac'}

\item{...}{for \code{\link{plot}}, arguments passed to the function 
\code{\link[stats]{plot.dendrogram}}. Default values for \code{type} and
\code{leaflab} are respectively set to \code{"triangle"} and \code{"none"}}

\item{y}{not used}

\item{mode}{type of dendrogram to plot (see Details). Default to 
\code{"standard"}}

\item{graph}{(logical) whether the diagnostic plot has to be displayed or 
not. Default to \code{TRUE}}

\item{verbose}{(logical) whether to print a summary of the result or not.
Default to \code{TRUE}}
}
\description{
Methods for class 'chac'
}
\details{
When \code{\link{plot.chac}} is called with 
\code{mode = "standard"}, the standard dendrogram is plotted, even though,
due to contingency constrains, some branches are reversed (decreasing
merges). When \code{\link{plot.chac}} is called with 
\code{mode = "corrected"}, a correction is applied to original heights so as
to have only non decreasing merges). It does not change the result of the 
clustering, only the look of the dendrogram for easier interpretation. When
\code{\link{plot.chac}} is called with \code{mode = "inertia"}, heights of
the dendrogram correspond to the cluster inertia instead of the cluster
linkage value. This also ensures non decreasing merges.

\code{\link{diagnose}} invisibly exports a data frame with the 
numbers of decreasing merges described by the labels of the clusters being
merged at this step and at the previous one, as well as the corresponding
merge heights.
}
