% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicclust(1).R
\name{hicclust}
\alias{hicclust}
\title{Constrained Hierarchical Agglomerative Clustering of Genomic regions(loci)}
\usage{
hicclust(x, h, ...)
}
\arguments{
\item{x}{x can be:
1. A pxp contact map of class Matrix::dsCMatrix in which the entries are the number of counts of physical interactions observed between all pairs of loci
2. An object of class HiTC::HTCexp. The corresponding Hi-C data is stored as a Matrix::dsCMatrix object in the intdata slot
3. A text file with one line per pair of loci for which an interaction has been observed (in the format: locus1<tab>locus2<tab>signal).}

\item{h}{band width. It is assumed that the similarity between two items is 0 when these 
items are at a distance of more than band width h}

\item{\dots}{further arguments to be passed to read.table function.It can contain any argument(s) of 
\code{\link{read.table}} except the file argument.}
}
\value{
Function \code{hicclust} returns an object of class \code{\link[stats]{hclust}}.
}
\description{
Function to perform adjacency-constrained hierarchical agglomerative clustering 
of genomic regions(loci)
}
\details{
\code{hicclust} performs constrained hierarchichal agglomerative clustering of 
genomic regions (loci) according to information provided by high-throughput conformation capture
 data (Hi-C). Constrained Hierarchical Agglomerative Clustering is hierarchical agglomerative 
 clustering in which each observation is associated to a position, and the clustering is 
 constrained so as only adjacent clusters are merged.
}
\examples{
#Input as HiTC::HTCexp object
library("HiTC")
data("hic_imr90_40$chrXchrX", package="adjclust")
h <- 3881
res1 <- hicclust(obj, h)

#Input as Matrix::dsCMatrix contact map
mat <- intdata(obj) 
res2 <- hicclust(mat, h)

#Input as text file
h <- 5
res3 <- hicclust(system.file("extdata", "sample.txt", package = "adjclust"), h)

}
