% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.100.R
\docType{data}
\name{R2.100}
\alias{R2.100}
\title{Linkage disequilibrium on chromosome 22 in the European population}
\format{A list of two numeric vectors of length \eqn{p-h*(h+1)/2}:
\describe{
  \item{R2.100}{"R squared" LD statistic}
  \item{Dprime.100}{"R squared" LD statistic}
}
\code{h} is the size of the diagonal band considered}
\source{
Data extracted from \code{snpStats::ld.example}. The original data is from the HapMap project.
}
\usage{
R2.100
}
\description{
A dataset containing linkage disequilibrium (LD) statistics ("R-squared" and "D
prime") for \eqn{p=603} SNPs spanning a one megabase regions on chromosome 22, in a 
sample of 90 Europeans (CEPH)
}
\keyword{datasets}

