% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpclust.R
\name{snpclust}
\alias{snpclust}
\title{Constrained Hierarchical Agglomerative Clustering of Single Nucleotide Polymorphisms}
\usage{
snpclust(x, h, stats)
}
\arguments{
\item{x}{either a genotype matrix of class snpStats::SnpMatrix/base::matrix or a linkage disequlibrium matrix of 
class Matrix::dgCMatrix}

\item{h}{band width. It is assumed that the similarity between two items is 0 when these items are at a 
distance of more than band width h}

\item{stats}{a character vector specifying the linkage disequilibrium measure to be calculated. This should contain one or 
more of the strings: "LLR", "OR", "Q", "Covar", "D.prime", "R.squared" or "R"}
}
\value{
Function \code{snpclust} returns an object of class \code{\link[stats]{hclust}}.
}
\description{
Function to perform adjacency-constrained hierarchical agglomerative clustering of Single Nucleotide Polymorphisms
}
\details{
\code{snpclust} performs constrained hierarchichal agglomerative clustering of single nucleotide 
polymorphisms. Constrained Hierarchical Agglomerative Clustering is hierarchical agglomerative
clustering in which each observation is associated to a position, and the clustering is 
constrained so as to merge only adjacent clusters.
}
\examples{
library(snpStats)
data(testdata)

#Input as snpStats::SnpMatrix
fit1 <- snpclust(Autosomes[1:200, 1:5], 3, "R.squared")

#Input as base::matrix
fit2 <- snpclust(as.matrix(Autosomes[1:200, 1:5]), 3, "R.squared")

#Input as Matrix::dgCMatrix
ld <- ld(Autosomes[1:200, 1:5], depth=3, stats="R.squared")
fit3 <- snpclust(ld, 3)

}
